Add files via upload
Drag the weapons you link to the characters hand into the weapon slots for the script. Also if it is not added drag the weapon switching and  equipping weapon script to the switch weapon menu prefab. Press Q to open the weapon wheel, and press Q to close the weapon wheel